Quite a few uncommitted changes:
* Spelling fixes

* artificialRating's. Using some awful theory, we parse the comments to
decide whether or not people enjoyed visiting the cache. It's not perfect
yet, but may be useful. Use -a and -A to decide min/max rating (1-5).
Inspired by a discussion with Dmitry Pruss.

* bugfix: setting terrainMax added difficultyMin flags to the filename and
title. Oops.

* "last" field in HTML and text now reads "last comment"

* details.rb cleanup. We now set the commenttypes pretty accurately, and
have merged the visitors table with that code. Moved html cleaning code to 
it's own subroutine, cleanHTML(), where it could use some real optimization.